+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed MSRV to 1.66.1.
 - Improved performance of `delete-orphaned-files` command.
 - Renamed `fetch-activity` command to `import-activity`.
+- Renamed `fetch-actor` command to `import-actor`.
 
 ### Fixed